Create sshZ.src
Requires port 22 on victum.
Gets Bank.txt, Mail.txt and passwd contents and adds them the home_dir + "/yoinked/" in folders of the same name.
Then it checks password bank for pass hash, if not then it deciphers it after deciphering it adds the pass with hash to password bank then connects to server ip as root.

Requires libssh.so v1.0.0